api/requirements.txt
altair==4.1.0
argh==0.26.2
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
boto3==1.18.62
botocore==1.21.62
branca==0.4.2
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==7.1.2
cryptography==35.0.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
dnspython==1.16.0
entrypoints==0.3
Flask==1.1.2
folium==0.11.0
fpdf==1.7.2
gitdb==4.0.7
GitPython==3.1.24
idna==3.1
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==4.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
olefile==0.46
packaging==21.0
pandas==1.2.5
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pip==21.3
plotly==5.3.1
plotly-express==0.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.2
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pymongo==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.2.0
six==1.16.0
smmap==3.0.5
soupsieve==2.0.1
statsmodels==0.12.2
streamlit==1.0.0
streamlit-folium==0.4.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
zipp==3.6.0
